Move get_check_env function to a common place
Move `get_check_env` function to a common place because this
function is used in the analyze and server part of the
CodeChecker too.
